+++ src/components/MypageContent.tsx
@@ -1,0 +1,31 @@
+"use client";
+
+import CategoryButton from "@/components/CategoryButton";
+import Tab from "@/components/Tab";
+
+const tabItems1 = ["나의 모임", "나의 리뷰", "내가 만든 모임"];
+const categories = ["작성 가능한 리뷰", "작성한 리뷰"];
+
+export default function MypageContent() {
+  return (
+    <div className="flex flex-1 flex-col gap-6 border-t-2 border-gray-900 bg-white px-6 pt-6">
+      <Tab
+        category={
+          <CategoryButton categories={categories}>
+            {categories.map((category) => (
+              <CategoryButton.Title key={category} category={category} />
+            ))}
+          </CategoryButton>
+        }
+        targetIndex={1}
+      >
+        {tabItems1.map((tabItem, idx) => (
+          <Tab.Item key={tabItem} index={idx}>
+            {tabItem}
+          </Tab.Item>
+        ))}
+      </Tab>
+      <div className="border">리스트 컨테이너</div>
+    </div>
+  );
+}
